backend/requirements.txt
fastapi
uvicorn
spacy
python-dotenv
google-genai
PyMuPDF
python-multipart
starlette
python-dateutil
